refactor: restructure Electron project directories and update build paths
