+++ forge.config.ts
@@ -20,17 +20,14 @@ const config: ForgeConfig = {
   ],
   plugins: [
     new VitePlugin({
-      // `build` can specify multiple entry builds, which can be Main process, Preload scripts, Worker process, etc.
-      // If you are familiar with Vite configuration, it will look really familiar.
       build: [
         {
-          // `entry` is just an alias for `build.lib.entry` in the corresponding file of `config`.
-          entry: "src/main.ts",
+          entry: "src/main/main.ts",
           config: "vite.main.config.ts",
           target: "main",
         },
         {
-          entry: "src/preload.ts",
+          entry: "src/preload/preload.ts",
           config: "vite.preload.config.ts",
           target: "preload",
         },
