+++ index.html
@@ -10,6 +10,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/renderer.tsx"></script>
+    <script type="module" src="/src/renderer/renderer.tsx"></script>
   </body>
 </html>
