+++ package.json
@@ -12,7 +12,8 @@
     "lint": "eslint src",
     "lint:fix": "eslint src --fix",
     "format": "prettier --write .",
-    "prepare": "husky"
+    "prepare": "husky",
+    "postinstall": "electron-rebuild -f -w better-sqlite3"
   },
   "lint-staged": {
     "*.{ts,tsx}": [
@@ -40,6 +41,7 @@
     "@electron-forge/plugin-vite": "^7.9.0",
     "@electron/fuses": "^1.8.0",
     "@tailwindcss/postcss": "^4.1.13",
+    "@types/better-sqlite3": "^7.6.13",
     "@types/electron-squirrel-startup": "^1.0.2",
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
@@ -62,6 +64,7 @@
     "vite": "^5.4.20"
   },
   "dependencies": {
+    "better-sqlite3": "^12.4.1",
     "clsx": "^2.1.1",
     "electron-squirrel-startup": "^1.0.1",
     "lucide-react": "^0.544.0",
