--- src/pages/index.ts
@@ -1,7 +1,0 @@
-import HistoryPage from "@/pages/history";
-import ShortcutsPage from "@/pages/shortcuts";
-import ModesPage from "@/pages/modes";
-import LoginPage from "@/pages/login";
-import NotFoundPage from "./NotFoundPage";
-
-export { HistoryPage, ShortcutsPage, ModesPage, LoginPage, NotFoundPage };
