+++ src/renderer/components/ModeBadges.tsx
