+++ src/renderer/hooks/useSplashController.ts
