+++ src/renderer/layouts/app-shell/header/AuthButton.tsx
@@ -1,6 +1,6 @@
 import { LogIn } from "lucide-react";
 import { useNavigate } from "react-router-dom";
-import ROUTES from "@/constants/routes";
+import ROUTES from "@/shared/constants/routes";
 
 function AuthButton() {
   const navigate = useNavigate();
