+++ src/renderer/pages/history/EmptyState.tsx
