+++ src/renderer/pages/history/details/Arrow.tsx
