+++ src/renderer/pages/history/details/MessageCard.tsx
