+++ src/renderer/pages/index.ts
@@ -1,0 +1,7 @@
+import HistoryPage from "@/renderer/pages/history";
+import ShortcutsPage from "@/renderer/pages/shortcuts";
+import ModesPage from "@/renderer/pages/modes";
+import LoginPage from "@/renderer/pages/login";
+import NotFoundPage from "./NotFoundPage";
+
+export { HistoryPage, ShortcutsPage, ModesPage, LoginPage, NotFoundPage };
