+++ src/renderer/pages/login/SkipLinkButton.tsx
