+++ src/renderer/pages/modes/index.tsx
