+++ src/renderer/pages/shortcuts/index.tsx
