+++ src/renderer/pages/splash/LoadingIndicator.tsx
