+++ src/renderer/pages/splash/LogoBox.tsx
