+++ src/renderer/pages/splash/ServiceTitle.tsx
