+++ src/renderer/pages/splash/VersionFooter.tsx
