+++ src/renderer/pages/splash/index.tsx
