+++ src/renderer/renderer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
+import "@/renderer/styles/index.css";
 import { MemoryRouter } from "react-router-dom";
 import App from "./App";
