+++ src/shared/types/images.d.ts
